requirements.txt
fastapi
uvicorn
transformers
torch
openai-whisper
pandas
python-multipart
python-dotenv
moviepy
opencv-python
deepface